Add service number
