+++ src/applications/pensions/tests/config/servicePeriods.unit.spec.jsx
@@ -7,6 +7,7 @@ import {
   DefinitionTester,
   getFormDOM,
 } from 'platform/testing/unit/schemaform-utils.jsx';
+import { waitFor } from '@testing-library/dom';
 import formConfig from '../../config/form';
 
 describe('Pensions service periods', () => {
@@ -98,7 +99,7 @@ describe('Pensions service periods', () => {
     expect(formDOM.querySelectorAll('.usa-alert').length).to.equal(0);
   });
 
-  it('should add another service period', () => {
+  it('should add another service period', async () => {
     const onSubmit = sinon.spy();
     const form = ReactTestUtils.renderIntoDocument(
       <DefinitionTester
@@ -126,13 +127,14 @@ describe('Pensions service periods', () => {
     ReactTestUtils.Simulate.click(
       formDOM.querySelector('.va-growable-add-btn'),
     );
-
-    expect(
-      formDOM.querySelector('.va-growable-background').textContent,
-    ).to.contain('Army');
+    waitFor(() => {
+      expect(
+        formDOM.querySelector('.va-growable-background').textContent,
+      ).to.contain('Army');
+    });
   });
 
-  it('should submit with valid data', () => {
+  it('should submit with valid data', async () => {
     const onSubmit = sinon.spy();
     const form = ReactTestUtils.renderIntoDocument(
       <DefinitionTester
@@ -156,8 +158,9 @@ describe('Pensions service periods', () => {
     );
 
     formDOM.submitForm();
-
-    expect(formDOM.querySelectorAll('.usa-input-error').length).to.equal(0);
-    expect(onSubmit.called).to.be.true;
+    waitFor(() => {
+      expect(formDOM.querySelectorAll('.usa-input-error').length).to.equal(0);
+      expect(onSubmit.called).to.be.true;
+    });
   });
 });
